refactor(component): refactor Modifiers for visibility
- 한 모듈 내에 너무 많은 코드가 존재하여 모듈을 기능별로 분리했다.
