+++ apps/web/templates/layouts/sidebar/right/border-modifier/Matrix.tsx
@@ -2,14 +2,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { useTheme } from '@emotion/react';
 
-import {
-  DefaultBox,
-  DefaultDivider,
-  DefaultHStack,
-  DefaultText,
-  DefaultVStack,
-  StrongText,
-} from 'ui';
+import { DefaultBox, DefaultHStack, DefaultText, DefaultVStack, StrongText } from 'ui';
 
 import {
   DirectionsContstants,
@@ -17,21 +10,7 @@ import {
   concurrentlyActivedSections,
 } from '@atoms/blockBorderAtom';
 
-import { useBorderMatrix, useBorderModifier } from '@hooks/index';
-
-import { TemplatedColorInputWithTitlePresenter } from './TemplatedColorInputWithTitle';
-import { TemplatedInputWithTitlePresenter } from './TemplatedInputWithTitlePresenter';
-
-const initialmatrixState = {
-  topLeft: false,
-  top: false,
-  topRight: false,
-  left: false,
-  right: false,
-  bottomLeft: false,
-  bottom: false,
-  bottomRight: false,
-};
+import { useBorderMatrix } from '@hooks/index';
 
 interface SubMatrixCommonPropsInterface {
   onMouseOverMatrix: (key: EdgeDirectionsContstants | DirectionsContstants | 'all') => void;
@@ -53,6 +32,17 @@ interface EdgeMatrixPropsInterface extends SubMatrixCommonPropsInterface {
   actived: boolean;
   position: EdgeDirectionsContstants;
 }
+
+const initialmatrixState = {
+  topLeft: false,
+  top: false,
+  topRight: false,
+  left: false,
+  right: false,
+  bottomLeft: false,
+  bottom: false,
+  bottomRight: false,
+};
 
 const EdgeMatrix = ({
   actived,
@@ -249,135 +239,5 @@ export function BorderMatrix() {
         </DefaultHStack>
       </DefaultVStack>
     </DefaultVStack>
-  );
-}
-
-/**
- * @description
- * INFO: 보더를 설정함에 있어 모서리일 경우랑 선의 경우를 다르게 인터페이스를 가져갈 필요가 있었다.
- * 이를 좀 더 유연하게 설계하기 위해 팩토리 메서드 패턴을 사용했다.
- */
-export function BorderSemiModifierFactory() {
-  const { blockBorderState } = useBorderMatrix();
-
-  return blockBorderState.activeBorder in EdgeDirectionsContstants ? (
-    <EdgeModifier />
-  ) : (
-    <LineModifier />
-  );
-}
-
-export function EdgeModifier() {
-  const { activeSectionBorderRadius, setBorderRadiusMiddleWare } = useBorderModifier();
-
-  return (
-    <DefaultVStack spacing={2} paddingTop="4px">
-      <DefaultHStack spacing={2}>
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="42px"
-          title="둥글기"
-          placeholder="입력"
-          value={activeSectionBorderRadius()}
-          onChange={setBorderRadiusMiddleWare}
-        />
-      </DefaultHStack>
-    </DefaultVStack>
-  );
-}
-
-export function LineModifier() {
-  const {
-    activeSectionBorderColor,
-    activeSectionBorderOpacity,
-    activeSectionBorderRadius,
-    activeSectionBorderStyle,
-    activeSectionBorderWidth,
-    setBorderMiddleware,
-    setBorderRadiusMiddleWare,
-  } = useBorderModifier();
-  return (
-    <DefaultVStack spacing={2} paddingTop="4px">
-      <DefaultHStack spacing={2}>
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="42px"
-          title="두께"
-          placeholder="입력"
-          value={activeSectionBorderWidth()}
-          onChange={(e) => setBorderMiddleware(e, 'width')}
-        />
-
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="42px"
-          title="둥글기"
-          placeholder="입력"
-          value={activeSectionBorderRadius()}
-          onChange={setBorderRadiusMiddleWare}
-        />
-
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="48px"
-          title="스타일"
-          placeholder="입력"
-          value={activeSectionBorderStyle()}
-          onChange={(e) => setBorderMiddleware(e, 'style')}
-        />
-      </DefaultHStack>
-
-      <DefaultHStack spacing={2}>
-        <TemplatedColorInputWithTitlePresenter
-          direction="vertical"
-          width="24px"
-          title="색상"
-          value={activeSectionBorderColor()}
-          onChange={(e) => setBorderMiddleware(e, 'color')}
-        />
-
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="60px"
-          title="색상번호"
-          placeholder="입력"
-          value={activeSectionBorderColor()}
-          onChange={(e) => setBorderMiddleware(e, 'color')}
-        />
-
-        <TemplatedInputWithTitlePresenter
-          direction="vertical"
-          inputWidth="48px"
-          title="투명도"
-          placeholder="입력"
-          value={activeSectionBorderOpacity()}
-          onChange={(e) => setBorderMiddleware(e, 'opacity')}
-        />
-      </DefaultHStack>
-    </DefaultVStack>
-  );
-}
-
-/* eslint-disable no-console */
-export function ActivedBlockBorderModifier() {
-  const theme = useTheme();
-  const { blockBorderState } = useBorderMatrix();
-
-  return (
-    <>
-      <DefaultVStack spacing={4}>
-        <StrongText size={theme.fontSize.sm} color="white">
-          블록 테두리({blockBorderState.name})
-        </StrongText>
-
-        <DefaultHStack spacing={2}>
-          <BorderMatrix />
-
-          <BorderSemiModifierFactory />
-        </DefaultHStack>
-      </DefaultVStack>
-
-      <DefaultDivider horizontal size="100%" borderColor="white" />
-    </>
   );
 }
